requirements.txt
pyyaml==5.4.1
pandas==1.2.2
requests==2.22.0